fix types, tidy up

+++ src/components/ApplicationQuestionsList/ApplicationQuestionsList.tsx
@@ -6,12 +6,11 @@ import Table from "@govuk-react/table";
 import Button from "@govuk-react/button";
 
 import { LinkButton } from "../../theme";
-import { ApplicationQuestion } from "../../types";
+import { UpdateApplicationQuestionInput } from "../../API";
 
 interface Props {
-    // questions: (ApplicationQuestion[] | null);
-    questions: any; //ApplicationQuestion[] | null;
     applicationId: string;
+    questions: (UpdateApplicationQuestionInput | null)[] | null;
     addQuestion: (id: string) => void;
     deleteQuestion: (id: string) => void;
 }
@@ -22,42 +21,44 @@ export const ApplicationQuestionsList: FC<Props> = ({
     addQuestion,
     deleteQuestion
 }) => {
-    console.log(questions);
-
     return (
         <>
             <Caption> Application questions </Caption>
             <Table>
                 {questions &&
-                    questions.map((question: ApplicationQuestion) => {
-                        return (
-                            <Table.Row key={question.id}>
-                                <Table.Cell>
-                                    <Link
-                                        to={`/component/application/${applicationId}/question/${
-                                            question.id
-                                        }`}
-                                    >
-                                        {question.heading}
-                                    </Link>
-                                </Table.Cell>
-                                <Table.Cell>
-                                    <LinkButton
-                                        onClick={() =>
-                                            deleteQuestion(question.id)
-                                        }
-                                    >
-                                        Remove
-                                    </LinkButton>
-                                </Table.Cell>
-                                <Table.Cell>
-                                    {question.complete
-                                        ? "Complete"
-                                        : "Incomplete"}
-                                </Table.Cell>
-                            </Table.Row>
-                        );
-                    })}
+                    questions.map(
+                        (question: UpdateApplicationQuestionInput | null) => {
+                            if (question !== null) {
+                                return (
+                                    <Table.Row key={question.id}>
+                                        <Table.Cell>
+                                            <Link
+                                                to={`/component/application/${applicationId}/question/${
+                                                    question.id
+                                                }`}
+                                            >
+                                                {question.heading}
+                                            </Link>
+                                        </Table.Cell>
+                                        <Table.Cell>
+                                            <LinkButton
+                                                onClick={() =>
+                                                    deleteQuestion(question.id)
+                                                }
+                                            >
+                                                Remove
+                                            </LinkButton>
+                                        </Table.Cell>
+                                        <Table.Cell>
+                                            {question.complete
+                                                ? "Complete"
+                                                : "Incomplete"}
+                                        </Table.Cell>
+                                    </Table.Row>
+                                );
+                            }
+                        }
+                    )}
                 <Table.Row>
                     <Table.Cell>
                         <Button onClick={() => addQuestion(applicationId)}>
